delete access cody in body

+++ public/src/pages/Edit.js
@@ -90,21 +90,26 @@ export default function Edit() {
 
   async function deletePosting(uuid){
     let api_url = api.gateway + "/postings/delete?uuid=" + uuid;
-
-    await fetch(api.gateway, {
-        method: 'DELETE', // *GET, POST, PUT, DELETE, etc.
-        mode: 'cors', // no-cors, *cors, same-origin
-        cache: 'no-cache', // *default, no-cache, reload, force-cache, only-if-cached
-        credentials: 'same-origin', // include, *same-origin, omit
-        headers: {
-          'Content-Type': 'application/json'
-        },
-        redirect: 'follow', // manual, *follow, error
-        referrerPolicy: 'no-referrer', // no-referrer, *no-referrer-when-downgrade, origin, origin-when-cross-origin, same-origin, strict-origin, strict-origin-when-cross-origin, unsafe-url
-    })
-    .then(() => {
-        alert("Post has been deleted.");
-    })
+    
+    if(accessCode){
+      await fetch(api.gateway, {
+          method: 'DELETE', // *GET, POST, PUT, DELETE, etc.
+          mode: 'cors', // no-cors, *cors, same-origin
+          cache: 'no-cache', // *default, no-cache, reload, force-cache, only-if-cached
+          credentials: 'same-origin', // include, *same-origin, omit
+          headers: {
+            'Content-Type': 'application/json'
+          },
+          redirect: 'follow', // manual, *follow, error
+          referrerPolicy: 'no-referrer', // no-referrer, *no-referrer-when-downgrade, origin, origin-when-cross-origin, same-origin, strict-origin, strict-origin-when-cross-origin, unsafe-url,
+          body: JSON.stringify(accessCode)
+      })
+      .then(response => response.json())
+      .then(data => {
+        console.log(data);
+        alert("Attempting to delete posting.");
+      })
+    }
   }
 
   async function getCaptchaToken() {
@@ -244,7 +249,7 @@ export default function Edit() {
               var postObj = checkMissingAttributes();
               // console.log("postobj", postObj);
               
-              if (postObj != null) {
+              if (postObj != null && accessCode) {
                   // postObj.token = await getCaptchaToken();
                   await postDetails(postObj);
                   alert("Post saved!");
@@ -260,12 +265,17 @@ export default function Edit() {
 
 
           <Button className="my-4 mr-4" variant="outline-danger" disabled={deleteBtnClicked} onClick={() => {
-              let deleteDecision = window.confirm("Are you sure you want to delete this post?");
-              if(deleteDecision){
-                deletePosting(id);
-                setDeleteBtnClicked(true);
-                window.location.href = "/";
-              }           
+              if(accessCode){
+                let deleteDecision = window.confirm("Are you sure you want to delete this post?");
+                if(deleteDecision){
+                  deletePosting(id);
+                  setDeleteBtnClicked(true);
+                  // window.location.href = "/";
+                }
+              }
+              else{
+                alert("Please enter a valid access code.");
+              }
           }}>
             Delete Posting
           </Button>
